requirement.txt
amqp==2.6.0
billiard==3.6.3.0
celery==4.4.3
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.4
django-haystack==2.8.1
django-redis==4.11.0
elasticsearch==2.4.1
fdfs-client-py==1.2.6
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.10
MarkupSafe==1.1.1
mutagen==1.44.0
mysqlclient==1.4.6
Pillow==7.1.2
pytz==2020.1
QQLoginTool==0.3.0
redis==3.5.3
requests==2.23.0
setuptools-scm==4.1.2
SQLAlchemy==1.3.17
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
